refactor(team-builder): extract shared JWT decode helpers
